Rename file system node metadata list accessor to match canonical declaration
