+++ packages/voictents-and-estinants-engine/src/adapted-programs/programmable-units/file/fileSystemObjectEnumeratorConfiguration.ts
@@ -1,7 +1,7 @@
 import {
   ComparisonConfigurationTypeName,
   FilePathAccessorInput,
-} from '../../../package-agnostic-utilities/file/getNestedFilePaths';
+} from '../../../package-agnostic-utilities/file/getNestedFileSystemNodeMetadataList';
 import { StandardInMemoryVoque } from '../../../layer-agnostic-utilities/voque/inMemoryVoque';
 
 /**
