+++ packages/voictents-and-estinants-engine/src/package-agnostic-utilities/file/getNestedFileSystemNodeMetadataList.ts
